@@ -1,6 +1,6 @@
 {
   "name": "input-material-ui",
-  "version": "2.9.0",
+  "version": "3.0.0",
   "description": "A material-ui INPUT component which allows to see/hide the password text of the field and to clear the input value",
   "main": "lib/index.js",
   "sideEffects": false,
@@ -11,7 +11,8 @@
     "storybook:start": "start-storybook -c storybook",
     "storybook:build": "build-storybook -c storybook -o docs",
     "storybook:commit": "git add . && git commit -m \"Updated storybook\"",
-    "all": "yarn install && yarn storybook:build && yarn deploy"
+    "all": "yarn reinstall && yarn storybook:build && yarn deploy",
+    "reinstall": "rimraf yarn.lock node_modules && yarn install"
   },
   "repository": {
     "type": "git",
@@ -39,45 +40,33 @@
   },
   "homepage": "https://github.com/iulian-radu-at/input-material-ui#readme",
   "dependencies": {
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/icons-material": "^5.10.16",
+    "@mui/material": "^5.10.17",
     "lodash": "^4.17.21"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.5",
-    "@storybook/react": "6.3.8",
+    "@babel/core": "^7.20.5",
+    "@storybook/builder-webpack5": "^6.5.14",
+    "@storybook/manager-webpack5": "^6.5.14",
+    "@storybook/react": "^6.5.14",
     "@types/babel-core": "^6.25.7",
-    "@types/enzyme": "^3.10.9",
-    "@types/enzyme-adapter-react-16": "^1.0.6",
-    "@types/lodash": "^4.14.172",
-    "@types/react": "^16.14.14",
-    "@types/react-dom": "^16.9.14",
+    "@types/lodash": "^4.14.191",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9",
     "@types/rimraf": "^3.0.2",
     "babel-core": "^6.26.3",
-    "babel-jest": "^27.1.0",
-    "babel-loader": "^8.2.2",
-    "babel-preset-react-app": "^10.0.0",
+    "babel-loader": "^9.1.0",
+    "babel-preset-react-app": "^10.0.1",
     "cross-env": "^7.0.3",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "rimraf": "^3.0.2",
-    "typescript": "^4.4.2"
+    "typescript": "^4.9.4"
   },
   "peerDependencies": {
-    "react": ">=16.9.0",
-    "react-dom": ">=16.9.0"
-  },
-  "jest": {
-    "modulePathIgnorePatterns": [
-      "./node_modules",
-      "./history"
-    ],
-    "transform": {
-      "^.+\\.js$": "babel-jest"
-    },
-    "testRegex": "src/*(/(__tests?__/)([^_].*/)*?[^_][^/]*?\\.(test|spec)?\\.(js?))$",
-    "setupFiles": [
-      "./enzymeAdapter.setup.js"
-    ]
+    "react": ">=18.0.0",
+    "react-dom": ">=18.0.0"
   }
 }
